refactor(step-3):패키지 경로 재설정
- util에 많은 책임이 있어, request, response 패키지를 생성해 나누었습니다.
