+++ src/main/java/request/HttpRequestUtils.java
@@ -1,4 +1,4 @@
-package util;
+package request;
 
 import java.util.HashMap;
 import java.util.Map;
