+++ src/test/java/controller/UserControllerTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import util.HttpRequest;
+import request.HttpRequest;
 
 class UserControllerTest {
     private UserController userController;
