+++ src/test/java/util/HttpRequestTest.java
@@ -4,6 +4,8 @@ import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import request.HttpRequest;
+
 class HttpRequestTest {
 
     @Test
